Extract common code for cassandra persistence test
